requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.5.1
mysql-connector-python==9.3.0
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
yarl==1.20.1
